requirements.txt
Django==1.11.3
olefile==0.44
Pillow==4.2.1
pytz==2017.2
wheel==0.24.
